requirements.txt
colored==2.3.1
Gooey==1.0.8.1
isort==8.0.0
pillow==12.1.1
psutil==7.2.2
pygtrie==2.5.0
six==1.17.0
wxPython==4.2.5
